requirements.txt
google-generativeai>=0.3.0
pandas>=1.3.0
pdf2image>=1.16.0
Pillow>=9.0.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
tqdm>=4.62.0
flask>=2.0.0
gunicorn>=20.1.0
